feat: Add UrlCreator context
The idea here is have a place to handle with the creation of an Url.
Here we can receive as dependency injection the fn to generate the Slug, with it possible change how we would like to generate and how
"stub" the slug to test the unique slug on the database.

Here in the nexts commits, I can check if the URL is valid or not and even ask to generate a new slug when a conflict happen.