feat: Add URL validation
After the happy path be structered, we can improve incrementally. Now I've added the URL validation,
I don't know if this is the best URL validation, but I believe this can works.

A good idea to improve in the future, is check if the URL received by the user has the http or https, and if don't, add it.